@@ -6,10 +6,10 @@ repository: https://github.com/TatsuUkraine/dart_environment_config
 issue_tracker: https://github.com/TatsuUkraine/dart_environment_config/issues
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  dart_style: ">=1.0.0 <2.0.0"
-  code_builder: ">=3.0.0 <4.0.0"
-  args: ">=1.0.0 <2.0.0"
-  yaml: ">=2.0.0 <3.0.0"
+  dart_style: ^2.0.0
+  code_builder: ^4.0.0
+  args: ^2.0.0
+  yaml: ^3.1.0
