@@ -2,7 +2,7 @@ import 'config_error.dart';
 
 class ValidationError extends ConfigError {
   final String field;
-  final String message;
+  final String? message;
 
   ValidationError(this.field, [this.message]);
 
