@@ -5,9 +5,9 @@ class PlatformValueProvider {
   final Map<String, String> data = Platform.environment;
 
   /// Returns value from platform env variable
-  String getValue(String key) {
+  String? getValue(String? key) {
     if (data.containsKey(key)) {
-      return data[key];
+      return data[key!];
     }
 
     return null;
